% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_dataset}
\alias{ons_dataset}
\title{Get ONS dataset details}
\usage{
ons_dataset(id)
}
\arguments{
\item{id}{Dataset ID (e.g. "wellbeing-quarterly", "cpih01")}
}
\value{
tibble: edition, state, links_self, links_latest_version
}
\description{
Returns metadata for a specific dataset including available editions.
}
