% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons.R
\name{ons_datasets}
\alias{ons_datasets}
\title{List ONS datasets}
\usage{
ons_datasets(limit = 20, offset = 0)
}
\arguments{
\item{limit}{Maximum number of datasets to return (default 20, max 1000)}

\item{offset}{Offset for pagination (default 0)}
}
\value{
tibble: id, title, description, release_frequency, state,
last_updated, unit_of_measure
}
\description{
Returns available datasets from the ONS API.
}
